requirements.txt
pygame>=2.5.0
moviepy>=1.0.3
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0
